Move inthash code to separate file
