+++ lammps/src/COLVARS/inthash.cpp
@@ -1,0 +1,166 @@
+#include <cstring>
+#include <memory>
+
+#include "inthash.h"
+
+
+namespace IntHash_NS {
+
+/************************************************************************
+ * integer hash code:
+ ************************************************************************/
+
+/* inthash() - Hash function returns a hash number for a given key.
+ * tptr: Pointer to a hash table, key: The key to create a hash number for */
+int inthash(const inthash_t *tptr, int key) {
+  int hashvalue;
+
+  hashvalue = (((key*1103515249)>>tptr->downshift) & tptr->mask);
+  if (hashvalue < 0) {
+    hashvalue = 0;
+  }
+
+  return hashvalue;
+}
+
+/*
+ *  rebuild_table_int() - Create new hash table when old one fills up.
+ *
+ *  tptr: Pointer to a hash table
+ */
+void rebuild_table_int(inthash_t *tptr) {
+  inthash_node_t **old_bucket, *old_hash, *tmp;
+  int old_size, h, i;
+
+  old_bucket=tptr->bucket;
+  old_size=tptr->size;
+
+  /* create a new table and rehash old buckets */
+  inthash_init(tptr, old_size<<1);
+  for (i=0; i<old_size; i++) {
+    old_hash=old_bucket[i];
+    while (old_hash) {
+      tmp=old_hash;
+      old_hash=old_hash->next;
+      h=inthash(tptr, tmp->key);
+      tmp->next=tptr->bucket[h];
+      tptr->bucket[h]=tmp;
+      tptr->entries++;
+    } /* while */
+  } /* for */
+
+  /* free memory used by old table */
+  free(old_bucket);
+}
+
+/*
+ *  inthash_init() - Initialize a new hash table.
+ *
+ *  tptr: Pointer to the hash table to initialize
+ *  buckets: The number of initial buckets to create
+ */
+void inthash_init(inthash_t *tptr, int buckets) {
+
+  /* make sure we allocate something */
+  if (buckets==0)
+    buckets=16;
+
+  /* initialize the table */
+  tptr->entries=0;
+  tptr->size=2;
+  tptr->mask=1;
+  tptr->downshift=29;
+
+  /* ensure buckets is a power of 2 */
+  while (tptr->size<buckets) {
+    tptr->size<<=1;
+    tptr->mask=(tptr->mask<<1)+1;
+    tptr->downshift--;
+  } /* while */
+
+  /* allocate memory for table */
+  tptr->bucket=(inthash_node_t **) calloc(tptr->size, sizeof(inthash_node_t *));
+}
+
+/*
+ *  inthash_lookup() - Lookup an entry in the hash table and return a pointer to
+ *    it or HASH_FAIL if it wasn't found.
+ *
+ *  tptr: Pointer to the hash table
+ *  key: The key to lookup
+ */
+int inthash_lookup(void *ptr, int key) {
+  const inthash_t *tptr = (const inthash_t *) ptr;
+  int h;
+  inthash_node_t *node;
+
+
+  /* find the entry in the hash table */
+  h=inthash(tptr, key);
+  for (node=tptr->bucket[h]; node!=nullptr; node=node->next) {
+    if (node->key == key)
+      break;
+  }
+
+  /* return the entry if it exists, or HASH_FAIL */
+  return(node ? node->data : HASH_FAIL);
+}
+
+/*
+ *  inthash_insert() - Insert an entry into the hash table.  If the entry already
+ *  exists return a pointer to it, otherwise return HASH_FAIL.
+ *
+ *  tptr: A pointer to the hash table
+ *  key: The key to insert into the hash table
+ *  data: A pointer to the data to insert into the hash table
+ */
+int inthash_insert(inthash_t *tptr, int key, int data) {
+  int tmp;
+  inthash_node_t *node;
+  int h;
+
+  /* check to see if the entry exists */
+  if ((tmp=inthash_lookup(reinterpret_cast<void *>(tptr), key)) != HASH_FAIL)
+    return(tmp);
+
+  /* expand the table if needed */
+  while (tptr->entries>=HASH_LIMIT*tptr->size)
+    rebuild_table_int(tptr);
+
+  /* insert the new entry */
+  h=inthash(tptr, key);
+  node=(struct inthash_node_t *) malloc(sizeof(inthash_node_t));
+  node->data=data;
+  node->key=key;
+  node->next=tptr->bucket[h];
+  tptr->bucket[h]=node;
+  tptr->entries++;
+
+  return HASH_FAIL;
+}
+
+/*
+ * inthash_destroy() - Delete the entire table, and all remaining entries.
+ *
+ */
+void inthash_destroy(inthash_t *tptr) {
+  inthash_node_t *node, *last;
+  int i;
+
+  for (i=0; i<tptr->size; i++) {
+    node = tptr->bucket[i];
+    while (node != nullptr) {
+      last = node;
+      node = node->next;
+      free(last);
+    }
+  }
+
+  /* free the entire array of buckets */
+  if (tptr->bucket != nullptr) {
+    free(tptr->bucket);
+    memset(tptr, 0, sizeof(inthash_t));
+  }
+}
+
+}
